finished refactor: ' :
