--- GamePortal/src/frontend/actions/group_actions.js
@@ -1,5 +1,0 @@
-export function resetGroups() {
-    return {
-        type: 'RESET_GROUPS'
-    }
-}
