--- GamePortal/src/frontend/reducers/screen_reducer.js
@@ -1,16 +1,0 @@
-import { initialState } from '../store/initial_state'
-
-export const screenReducer = (state = initialState.screen, action) => {
-    switch (action.type) {
-        case 'LOADING':
-            return Object.assign({}, state, {
-                loading: action.loading
-            });
-        case 'SWITCH_SCREEN':
-            return Object.assign({}, state, {
-                screen: action.screen
-            });
-        default:
-            return state;
-    }
-};
