+++ GamePortal/src/resources/img/login_buttons/email.png
