+++ GamePortal/src/resources/styles/common_style.js
