+++ GamePortal/src/screens/game_renderer/component.js
@@ -8,7 +8,7 @@ import {
 } from 'react-native';
 
 import * as firebase from 'firebase';
-import styles from "../../frontend/styles/common_style";
+import styles from "../../resources/styles/common_style";
 
 export default class GameRendererComponent extends Component {
