+++ GamePortal/src/screens/login/default/component_ios.js
@@ -1,9 +1,9 @@
 import React, {Component} from 'react';
-import styles from "../../../frontend/styles/common_style";
+import styles from "../../../resources/styles/common_style";
 import {Image, Text, TouchableOpacity, View} from "react-native";
 import * as firebase from "firebase/index";
 import {AccessToken, LoginManager} from "react-native-fbsdk";
-import {firebaseConnect, firebaseLoginFlow} from "../../../backend/users/login";
+import {firebaseConnect, firebaseLoginFlow} from "../../../util/firebase_login";
 
 export default class iOSLoginDefaultComponent extends Component {
 
@@ -46,7 +46,7 @@ export default class iOSLoginDefaultComponent extends Component {
         return (
             <TouchableOpacity onPress={() => this.facebookLogin()}>
                 <Image
-                    source={require('../../../frontend/img/login_buttons/facebook.png')}
+                    source={require('../../../resources/img/login_buttons/facebook.png')}
                 />
             </TouchableOpacity>
         );
@@ -58,7 +58,7 @@ export default class iOSLoginDefaultComponent extends Component {
         return (
             <TouchableOpacity onPress={() => switchScreen('LoginEmail')}>
                 <Image
-                    source={require('../../../frontend/img/login_buttons/email.png')}
+                    source={require('../../../resources/img/login_buttons/email.png')}
                 />
             </TouchableOpacity>
         );
@@ -68,7 +68,7 @@ export default class iOSLoginDefaultComponent extends Component {
         return (
             <TouchableOpacity onPress={() => this.anonymousLogin()}>
                 <Image
-                    source={require('../../../frontend/img/login_buttons/anonymous.png')}
+                    source={require('../../../resources/img/login_buttons/anonymous.png')}
                 />
             </TouchableOpacity>
         );
