+++ GamePortal/src/util/firebase_groups.js
