+++ GamePortal/src/util/firebase_login.js
