+++ GamePortal/src/util/firebase_public_fields.js
